+++ common/build.gradle.kts
@@ -11,6 +11,7 @@ kotlin {
     }
     androidMain.dependencies {
       implementation(libs.androidx.core)
+      implementation(libs.androidx.browser)
     }
   }
 }
