+++ ui/crash/src/androidMain/kotlin/com/github/kr328/clash/crash/ui/ApkBrokenScreen.kt
@@ -9,9 +9,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewWrapper
+import com.github.kr328.clash.common.util.TABBY_GITHUB
+import com.github.kr328.clash.common.util.openLink
 import com.github.kr328.clash.crash.R
-import com.github.kr328.clash.glue.util.TABBY_GITHUB
-import com.github.kr328.clash.glue.util.openLink
 import com.github.kr328.clash.ui.component.TabbyScaffold
 import com.github.kr328.clash.ui.icon.OutlineInfo
 import com.github.kr328.clash.ui.icon.TabbyIcons
