+++ ui/home/src/commonMain/kotlin/com/github/kr328/clash/home/vm/HelpViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.kr328.clash.common.di.AppInfoProvider
 import com.github.kr328.clash.common.log.Log
+import com.github.kr328.clash.common.util.TABBY_RELEASES_LATEST
 import com.github.kr328.clash.core.bridge.Bridge
-import com.github.kr328.clash.glue.util.TABBY_RELEASES_LATEST
 import com.github.kr328.clash.home.Res
 import com.github.kr328.clash.home.already_up_to_date
 import com.github.kr328.clash.home.api.HelpApi
